Ship Health values 0.0 from 1.0

+++ src/model/Machinegun.h
@@ -15,7 +15,7 @@ public:
 
     Machinegun();
 
-    int GetDamageValue() override;
+    float GetDamageValue() override;
 
 protected:
     void InnerInit() override;
